Do not disable vfat by default
On UEFI-systems the boot-partition is FAT by default (see [here](https://wiki.archlinux.org/index.php/Unified_Extensible_Firmware_Interface/System_partition)).

If we disable vfat, these systems become unbootable. This has already bitten some users using ansible-os-hardening (https://github.com/dev-sec/ansible-os-hardening/issues/162, https://github.com/dev-sec/ansible-os-hardening/issues/145).

Therefore I propose we do not check for a disabled vfat filesystem, if efi is used on these systems

+++ controls/os_spec.rb
@@ -210,7 +210,11 @@ control 'os-10' do
     its(:content) { should match 'install hfsplus /bin/true' }
     its(:content) { should match 'install squashfs /bin/true' }
     its(:content) { should match 'install udf /bin/true' }
-    its(:content) { should match 'install vfat /bin/true' }
+    # if efi is active, do not disable vfat. otherwise the system
+    # won't boot anymore
+    unless Dir.exist?('/sys/firmware/efi')
+      its(:content) { should match 'install vfat /bin/true' }
+    end
   end
 end
